+++ app/lib/database/period_database.dart
@@ -1,7 +1,7 @@
 import 'package:sqflite/sqflite.dart';
 import 'package:path/path.dart';
 import 'package:menstrudel/models/period_logs.dart';
-import 'package:menstrudel/models/period.dart';
+import 'package:menstrudel/models/periods/period.dart';
 
 class PeriodDatabase {
     static final PeriodDatabase instance = PeriodDatabase._init();
