+++ app/lib/screens/analytics_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:menstrudel/models/cycle_stats.dart';
 import 'package:menstrudel/models/monthly_cycle_data.dart';
-import 'package:menstrudel/models/period_stats.dart';
+import 'package:menstrudel/models/periods/period_stats.dart';
 import 'package:menstrudel/widgets/monthly_cycle_list_view.dart';
 import 'package:menstrudel/utils/period_predictor.dart';
 import 'package:menstrudel/database/period_database.dart';
