+++ app/lib/utils/period_predictor.dart
@@ -1,8 +1,8 @@
 import 'package:menstrudel/models/period_logs.dart';
-import 'package:menstrudel/models/period.dart';
+import 'package:menstrudel/models/periods/period.dart';
 import 'package:menstrudel/models/period_prediction_result.dart';
 import 'package:menstrudel/models/cycle_stats.dart';
-import 'package:menstrudel/models/period_stats.dart';
+import 'package:menstrudel/models/periods/period_stats.dart';
 import 'package:menstrudel/models/monthly_cycle_data.dart';
 import 'dart:math';
